Intermediate commit

+++ dockerwatch.sh
@@ -1,62 +1,35 @@
 #!/bin/bash
-
-# Check processes inside container
-# If "sacred process" is not running, then stop container;
-# If stop_dockerwatch string found inside sacred_proc_file, then exit dockerwatch.
 #
-# Created by Bryzgalov Peter on 2014/01/31.
-# Copyright (c) 2013 Riken AICS. All rights reserved
+# Check connection counter
+# If equal to 0, then stop container.
+#
+# Created by Bryzgalov Peter on 2014/02/20
+# Copyright (c) 2013-2014 Riken AICS. All rights reserved
 
-version="1.46"
+version="2.11"
 
-# Sacred processes read from file
-sacred_proc_file="/tmp/proc.tmp"
-# Store list of processes in tmp file
-tmp_file="/tmp/ps.tmp"
-# If this stinrg found in sacred_proc_file, then exit dockerwatch
-stop_dockerwatch="stopdockerwatch"
-timeout=3
+# Connections counter
+counter_file="/tmp/connection_counter"
+timeout=10
 
-ps_command="ps ax"
-
-function procsCheck {
-    # Check sacred process in running processes
-
-    $ps_command
-    $ps_command > $tmp_file
-    sacred_proc=$(cat $sacred_proc_file)
-    if grep -q "$sacred_proc" $tmp_file
-    then
-        echo "Matched"
-        return 1
-    else
-        echo "Sacred process not found in $tmp_file"
-    fi
-    # Check stop_dockerwatch string inside sacred_proc_file
-    if grep -q "$stop_dockerwatch" $sacred_proc_file
-    then
-        echo "Exit dockerwatch command found in $sacred_proc_file. Exiting."
-        exit 0
-    else
-        echo "Exit dockerwatch command not found in in $sacred_proc_file."
-    fi
-    return 0
+function synchro_read {
+    counter_file=$1
+    exec 20>$counter_file
+    flock -x -w 2 20
+    COUNTER=$(cat $counter_file)
+    return $COUNTER
 }
 
-echo "Start dockerwatch $version. Sacred process: $(cat $sacred_proc_file)"
+echo "Start dockerwatch $version"
 echo "--------------------------------------"
-procsCheck
-matched=$?
-
-while [ $matched -eq 1 ]
-do
-    sleep $timeout
-    # Check value in sacred proc file
-    procsCheck
-    matched=$?
-done
-
-# Stop container
-echo "Stopping container"
-echo "------------------"
-kill 1
+sleep $timeout
+synchro_read counter_file
+COUNTER=$?
+echo "counter = $COUNTER"
+if [ $COUNTER -eq "0" ]
+then
+    # Stop container
+    echo "Stopping container"
+    echo "------------------"
+    kill 1
+fi
